Reorganize action-generating code
