+++ src/map-interface/reducers/core.ts
