+++ src/map-interface/reducers/handlers/fetch.ts
